// src/module/Master/Promocodes/slice/Promocodes.type.ts
export interface Ipromocodes {
  pid: number;
  promoName: string;
  clientName: string;
  isAuth: number;
  firstName: string;
  lastName: string;
  city: string;
  town: string;
  state: string;
  zipCode: number;
  officePhone: number;
  personalPhone: number;
  payeeName: string;
  payPalEmailId: string;
  officeAdd: string;
  country: string;
}
export interface ICountry {
  tID: number;
  countryCode: string;
  countryName: string;
}